+++ extensions/omni_worker/omni_worker.hpp
@@ -3,3 +3,5 @@
 std::string arena_name();
 std::size_t arena_size();
 std::string mq_name();
+
+extern cppgres::worker *main_worker;