+++ extensions/omni_worker/src/instantiate.sql
@@ -18,7 +18,9 @@ begin
     insert
     into
         handlers
-    values ('MODULE_PATHNAME', 'sql');
+    values
+        ('MODULE_PATHNAME', 'sql'),
+        ('MODULE_PATHNAME', 'timer');
 
     create function reload_handlers() returns trigger
         language c as
@@ -39,6 +41,25 @@ begin
         stmt     text not null,
         position int  not null
     );
+
+    perform from pg_roles where rolname = 'omni_worker_timer_user';
+    if not found then
+        create role omni_worker_timer_user;
+    end if;
+
+    create function timer_after(delay_ms bigint, fun regproc,
+                                run_as regrole default current_role::regrole) returns bigint
+        language c as
+    'MODULE_PATHNAME';
+
+    create function timer_cancel(timer_id bigint) returns bool
+        language c as
+    'MODULE_PATHNAME';
+
+    revoke execute on function timer_after, timer_cancel from public, current_user;
+    grant execute on function timer_after, timer_cancel to omni_worker_timer_user;
+    execute format('grant all on schema %I to omni_worker_timer_user', schema);
+
 
 end;
 $$;